clean up code, refactor
