+++ src/containers/studies/StudiesReducer.js
@@ -2,25 +2,39 @@
  * @flow
  */
 
-import { Map, fromJS, List } from 'immutable';
+import { List, Map, fromJS } from 'immutable';
 import { RequestStates } from 'redux-reqseq';
 import type { SequenceAction } from 'redux-reqseq';
 
 import {
   ADD_PARTICIPANT,
+  CREATE_PARTICIPANTS_ENTITY_SET,
   CREATE_STUDY,
   GET_STUDIES,
   addStudyParticipant,
+  createParticipantsEntitySet,
   createStudy,
   getStudies,
 } from './StudiesActions';
 
+import { PROPERTY_TYPE_FQNS } from '../../core/edm/constants/FullyQualifiedNames';
 import { RESET_REQUEST_STATE } from '../../core/redux/ReduxActions';
 
+const { STUDY_ID } = PROPERTY_TYPE_FQNS;
+
 const INITIAL_STATE :Map<*, *> = fromJS({
-  [ADD_PARTICIPANT]: { requestState: RequestStates.STANDBY },
-  [CREATE_STUDY]: { requestState: RequestStates.STANDBY },
-  [GET_STUDIES]: { requestState: RequestStates.STANDBY },
+  [ADD_PARTICIPANT]: {
+    requestState: RequestStates.STANDBY
+  },
+  [CREATE_STUDY]: {
+    requestState: RequestStates.STANDBY
+  },
+  [CREATE_PARTICIPANTS_ENTITY_SET]: {
+    requestState: RequestStates.STANDBY
+  },
+  [GET_STUDIES]: {
+    requestState: RequestStates.STANDBY
+  },
 
   participants: Map(),
   studies: Map(),
@@ -54,8 +68,9 @@ export default function studiesReducer(state :Map<*, *> = INITIAL_STATE, action 
         REQUEST: () => state.setIn([CREATE_STUDY, 'requestState'], RequestStates.PENDING),
         FAILURE: () => state.setIn([CREATE_STUDY, 'requestState'], RequestStates.FAILURE),
         SUCCESS: () => {
-          const { studyUUID, study } = seqAction.value;
-          const updatedStudies = state.get('studies').set(studyUUID, fromJS(study));
+          const { study } = seqAction.value;
+          const studyId = study.getIn([STUDY_ID, 0]);
+          const updatedStudies = state.get('studies').set(studyId, fromJS(study));
 
           return state
             .set('studies', updatedStudies)
@@ -78,6 +93,21 @@ export default function studiesReducer(state :Map<*, *> = INITIAL_STATE, action 
           return state
             .setIn([ADD_PARTICIPANT, 'requestState'], RequestStates.SUCCESS)
             .setIn(['participants', studyId], participants);
+        }
+      });
+    }
+
+    case createParticipantsEntitySet.case(action.type): {
+      const seqAction :SequenceAction = action;
+      return createParticipantsEntitySet.reducer(state, action, {
+        REQUEST: () => state.setIn([CREATE_PARTICIPANTS_ENTITY_SET, 'requestState'], RequestStates.PENDING),
+        FAILURE: () => state.setIn([CREATE_PARTICIPANTS_ENTITY_SET, 'requestState'], RequestStates.FAILURE),
+        SUCCESS: () => {
+          const { entitySetName, entitySetId } = seqAction.value;
+          const updatedMap = state.get('participantEntitySetIds').set(entitySetName, entitySetId);
+          return state
+            .setIn([CREATE_PARTICIPANTS_ENTITY_SET, 'requestState'], RequestStates.SUCCESS)
+            .set('participantEntitySetIds', updatedMap);
         }
       });
     }
