+++ src/core/sagas/Sagas.js
@@ -27,7 +27,6 @@ export default function* sagas() :Generator<*, *, *> {
     // EDMSagas
     fork(EDMSagas.getEntityDataModelTypesWatcher),
     fork(EDMSagas.getAllEntitySetIdsWatcher),
-    fork(EDMSagas.createParticipantsEntitySetWatcher),
 
     // RoutingSagas
     fork(RoutingSagas.goToRootWatcher),
@@ -37,6 +36,7 @@ export default function* sagas() :Generator<*, *, *> {
     fork(StudiesSagas.getStudiesWatcher),
     fork(StudiesSagas.addStudyParticipantWatcher),
     fork(StudiesSagas.createStudyWatcher),
+    fork(StudiesSagas.createParticipantsEntitySetWatcher),
 
     // DataSagas
     fork(DataSagas.submitDataGraphWatcher)
